Unify toasts and notifications
