--- src/new-components/notifications/index.ts
@@ -1,2 +1,0 @@
-export { Notification, NotificationsProvider, useNotifications } from './notifications'
-export type { NotificationProps, NotificationsProviderProps } from './notifications'
