+++ src/new-components/toast/index.ts
@@ -1,2 +1,5 @@
-export { Toast } from './toast'
-export type { ToastProps } from './toast'
+export { StaticToast } from './static-toast'
+export { Toast, ToastsProvider, useToasts } from './use-toasts'
+
+export type { StaticToastProps } from './static-toast'
+export type { ToastProps, ToastsProviderProps } from './use-toasts'
