+++ src/new-components/toast/toast.stories.tsx
@@ -1,24 +1,119 @@
 import * as React from 'react'
 import { action as storybookAction } from '@storybook/addon-actions'
+
+import { AlertIcon } from '../icons/alert-icon'
+import { PasswordVisibleIcon } from '../icons/password-visible-icon'
+
+import { Box } from '../box'
 import { Button, ButtonProps } from '../button'
 import { CheckboxField } from '../checkbox-field'
 import { Heading } from '../heading'
-import { AlertIcon } from '../icons/alert-icon'
 import { Inline } from '../inline'
 import { SelectField } from '../select-field'
-import { Stack } from '../stack'
-import { Toast, ToastProps } from './toast'
 import { Spinner } from '../spinner'
-import { PasswordVisibleIcon } from '../icons/password-visible-icon'
+import { Stack } from '../stack'
+import { SwitchField } from '../switch-field'
+
+import { StaticToast, StaticToastProps } from './static-toast'
+import { Toast, ToastsProvider, useToasts } from './use-toasts'
+import { Text } from '../text'
 
 export default {
     title: 'Design system/Toast',
     parameters: {
         badges: ['accessible'],
+        layout: 'fullscreen',
     },
+    decorators: [
+        (Story: () => JSX.Element) => (
+            <ToastsProvider>
+                <Story />
+            </ToastsProvider>
+        ),
+    ],
 }
 
-export function ToastsStory() {
+const message = [
+    'Task was successfully deleted',
+    'Message was sent',
+    'Someone invited you to collaborate on a project',
+    'The project could not be deleted because there are other members collaborating on it',
+    'Your invitation could not be sent. Please try again.',
+    'There was an error trying to delete the channel. Please, try again.',
+    'An unknown error happened. If the issue persist, contact our support team.',
+]
+
+const actions = ['Undo', 'Retry', 'Details']
+
+function getRandom<T>(list: Array<T>): T {
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    return list[Math.floor(Math.random() * list.length)]!
+}
+
+export function NotificationToastsStory() {
+    const showToast = useToasts()
+    const count = React.useRef(0)
+    const [showSticky, setShowSticky] = React.useState(false)
+    return (
+        <Box padding="large">
+            <Stack space="xlarge">
+                <Heading level={1} size="larger">
+                    Toasts
+                </Heading>
+                <Text>
+                    Use the <code>useToast</code> hook to fire notification-like toasts.
+                </Text>
+                <Box>
+                    <Inline space="large">
+                        <Button
+                            variant="primary"
+                            onClick={() =>
+                                showToast({ message: `${count.current++}: ${getRandom(message)}` })
+                            }
+                        >
+                            Show toast
+                        </Button>
+                        <Button
+                            variant="primary"
+                            onClick={() => {
+                                const actionLabel = getRandom(actions)
+                                showToast({
+                                    message: `${count.current++}: ${getRandom(message)}`,
+                                    action: {
+                                        label: actionLabel,
+                                        onClick: storybookAction(actionLabel),
+                                    },
+                                })
+                            }}
+                        >
+                            Show toast with action
+                        </Button>
+                    </Inline>
+                </Box>
+                <SwitchField
+                    label="Show sticky toast?"
+                    checked={showSticky}
+                    onChange={(event) => setShowSticky(event.target.checked)}
+                    hint={
+                        <>
+                            Using the <code>&lt;Toast/&gt;</code> component, you can achieve having
+                            a toast that remains in view, and you control when it goes away.
+                        </>
+                    }
+                />
+            </Stack>
+            {showSticky ? (
+                <Toast
+                    message="This is a sticky toast"
+                    showDismissButton={false}
+                    autoDismissDelay={false}
+                />
+            ) : null}
+        </Box>
+    )
+}
+
+export function StaticToastStory() {
     const [showIcon, setShowIcon] = React.useState(false)
     const [showDismiss, setShowDismiss] = React.useState(false)
     const [actionVariant, setActionVariant] = React.useState<
@@ -26,7 +121,7 @@ export function ToastsStory() {
     >('none')
 
     const onClick = storybookAction('Toast action')
-    const action: ToastProps['action'] =
+    const action: StaticToastProps['action'] =
         actionVariant === 'default' ? (
             { label: 'Accept invitation', onClick }
         ) : actionVariant === 'loading' ? (
@@ -52,19 +147,27 @@ export function ToastsStory() {
     }
 
     return (
-        <Stack space="xxlarge">
-            <Inline space="large">
-                <CheckboxField
-                    label="Show icon?"
-                    checked={showIcon}
-                    onChange={(event) => setShowIcon(event.target.checked)}
-                />
-                <CheckboxField
-                    label="Show dismiss button?"
-                    checked={showDismiss}
-                    onChange={(event) => setShowDismiss(event.target.checked)}
-                />
-            </Inline>
+        <Stack space="xlarge" padding="large">
+            <Heading level={1} size="larger">
+                Statically-rendered toasts
+            </Heading>
+            <Text>
+                Use the <code>StaticToast</code> component to render a toast in custom positions.
+            </Text>
+            <Box>
+                <Inline space="large">
+                    <CheckboxField
+                        label="Show icon?"
+                        checked={showIcon}
+                        onChange={(event) => setShowIcon(event.target.checked)}
+                    />
+                    <CheckboxField
+                        label="Show dismiss button?"
+                        checked={showDismiss}
+                        onChange={(event) => setShowDismiss(event.target.checked)}
+                    />
+                </Inline>
+            </Box>
             <SelectField
                 label="Action button variant"
                 value={actionVariant}
@@ -90,12 +193,15 @@ export function ToastsStory() {
 
             <Stack space="medium">
                 <Heading level="2">Message only</Heading>
-                <Toast message="Someone invited you to collaborate on a project" {...commonProps} />
+                <StaticToast
+                    message="Someone invited you to collaborate on a project"
+                    {...commonProps}
+                />
             </Stack>
 
             <Stack space="medium">
                 <Heading level="2">Message and description</Heading>
-                <Toast
+                <StaticToast
                     message="Someone invited you to collaborate on a project"
                     description="You need to accept the invitation in order to collaborate."
                     {...commonProps}
@@ -104,7 +210,7 @@ export function ToastsStory() {
 
             <Stack space="medium">
                 <Heading level="2">Very long content</Heading>
-                <Toast
+                <StaticToast
                     message="The project you were invited to could not be loaded into your workspace at this time"
                     description="Please, try to access the project again in a few minutes. If the problem persists, contact our support team"
                     {...commonProps}
